Fix header build
